Remove browser notifications on Pusher events
