+++ src/lib/browser/index.ts
@@ -1,6 +1,5 @@
 import { provider } from '../../lib/cloudinary/provider';
 import { Message } from '../../store/messages';
-import { mapNotification as mapNotificationUtils } from '../../components/notification/utils';
 
 const DEFAULT_HEADING = 'Chat message received';
 
@@ -16,11 +15,4 @@ export const send = (options: { body; heading; tag }) => {
 
 export function mapMessage(message: Message) {
   return { heading: DEFAULT_HEADING, body: 'You have received a message', tag: message.id };
-}
-
-export function mapNotification(notification: Notification) {
-  // XXX verify with UX team that we want to display the user's name in a mention and reply notification
-  const { body, id } = mapNotificationUtils(notification);
-
-  return { heading: DEFAULT_HEADING, body, tag: id };
 }
