+++ src/store/notifications/api.ts
@@ -4,8 +4,3 @@ export async function fetchNotifications(userId: string) {
   const response = await get<any>('/api/notifications/filter', { limit: 15 }, { userId });
   return response.body;
 }
-
-export async function fetchNotification(notificationId: string) {
-  const response = await get<any>(`/api/notifications/${notificationId}/enhance`);
-  return response.body;
-}
